Adding stats...

+++ Assets/Scripts/Unit/Abilities/AbilityAction.cs
@@ -5,12 +5,12 @@ public abstract class AbilityAction : UnitAction
 {
 
     [SerializeField]
-    private int m_Charge;
-    public int charge { get { return m_Charge; } protected set { m_Charge = value; } }
+    private float m_Charge;
+    public float charge { get { return m_Charge; } protected set { m_Charge = value; } }
 
     [SerializeField]
-    private int m_Cost;
-    public int cost { get { return m_Cost; } protected set { m_Cost = value; } }
+    private float m_Cost;
+    public float cost { get { return m_Cost; } protected set { m_Cost = value; } }
     
     public AbilityAction() : base()
     {
